requirements.txt
# AndroidWorld Python Dependencies
# Core dependencies for running Android automation tasks

# Web scraping and automation
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.8.0

# Additional utilities
psutil>=5.9.0
pyyaml>=6.0
click>=8.1.0
rich>=12.0.0

# Genymotion Cloud integration
requests>=2.28.0

# Vertex AI integration
google-cloud-aiplatform>=1.34.0
google-cloud-monitoring>=2.15.0
google-cloud-logging>=2.15.0
google-cloud-trace>=1.8.0

# OpenTelemetry for tracing
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-cloud-trace>=0.10b0
opentelemetry-instrumentation-requests>=0.40b0
opentelemetry-instrumentation-urllib3>=0.40b0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
requests>=2.28.0

# Agent system dependencies
# (These are built-in Python modules, no additional installation needed)
# - json: Built-in
# - time: Built-in  
# - logging: Built-in
# - subprocess: Built-in
# - os: Built-in
# - uuid: Built-in
# - random: Built-in
# - datetime: Built-in
# - csv: Built-in
# - pathlib: Built-in (Python 3.4+)
